usm: classification: split the functions and helpers to protocol-dedicated-files
