+++ pkg/network/ebpf/c/protocols/http-classification-defs.h
@@ -1,0 +1,8 @@
+#ifndef __HTTP_CLASSIFICATION_DEFS_H
+#define __HTTP_CLASSIFICATION_DEFS_H
+
+// The minimal HTTP response has 17 characters: HTTP/1.1 200 OK\r\n
+// The minimal HTTP request has 16 characters: GET x HTTP/1.1\r\n
+#define HTTP_MIN_SIZE 16
+
+#endif
