+++ pkg/network/ebpf/c/protocols/http-classification-helpers.h
@@ -1,0 +1,37 @@
+#ifndef __HTTP_HELPERS_H
+#define __HTTP_HELPERS_H
+
+#include "protocol-classification-common.h"
+
+// Checks if the given buffers start with `HTTP` prefix (represents a response) or starts with `<method> /` which represents
+// a request, where <method> is one of: GET, POST, PUT, DELETE, HEAD, OPTIONS, or PATCH.
+static __always_inline bool is_http(const char *buf, __u32 size) {
+    CHECK_PRELIMINARY_BUFFER_CONDITIONS(buf, size, HTTP_MIN_SIZE)
+
+#define HTTP "HTTP/"
+#define GET "GET /"
+#define POST "POST /"
+#define PUT "PUT /"
+#define DELETE "DELETE /"
+#define HEAD "HEAD /"
+#define OPTIONS1 "OPTIONS /"
+#define OPTIONS2 "OPTIONS *"
+#define PATCH "PATCH /"
+
+    // memcmp returns
+    // 0 when s1 == s2,
+    // !0 when s1 != s2.
+    bool http = !(bpf_memcmp(buf, HTTP, sizeof(HTTP)-1)
+        && bpf_memcmp(buf, GET, sizeof(GET)-1)
+        && bpf_memcmp(buf, POST, sizeof(POST)-1)
+        && bpf_memcmp(buf, PUT, sizeof(PUT)-1)
+        && bpf_memcmp(buf, DELETE, sizeof(DELETE)-1)
+        && bpf_memcmp(buf, HEAD, sizeof(HEAD)-1)
+        && bpf_memcmp(buf, OPTIONS1, sizeof(OPTIONS1)-1)
+        && bpf_memcmp(buf, OPTIONS2, sizeof(OPTIONS2)-1)
+        && bpf_memcmp(buf, PATCH, sizeof(PATCH)-1));
+
+    return http;
+}
+
+#endif
