+++ pkg/network/ebpf/c/protocols/http2-defs.h
@@ -1,0 +1,38 @@
+#ifndef __HTTP2_DEFS_H
+#define __HTTP2_DEFS_H
+
+#include <linux/types.h>
+
+// Checkout https://datatracker.ietf.org/doc/html/rfc7540 under "HTTP/2 Connection Preface" section
+#define HTTP2_MARKER_SIZE 24
+
+// Checkout https://datatracker.ietf.org/doc/html/rfc7540 under "Frame Format" section
+#define HTTP2_FRAME_HEADER_SIZE 9
+#define HTTP2_SETTINGS_SIZE 6
+
+// All types of http2 frames exist in the protocol.
+// Checkout https://datatracker.ietf.org/doc/html/rfc7540 under "Frame Type Registry" section.
+typedef enum {
+    kDataFrame          = 0,
+    kHeadersFrame       = 1,
+    kPriorityFrame      = 2,
+    kRSTStreamFrame     = 3,
+    kSettingsFrame      = 4,
+    kPushPromiseFrame   = 5,
+    kPingFrame          = 6,
+    kGoAwayFrame        = 7,
+    kWindowUpdateFrame  = 8,
+    kContinuationFrame  = 9,
+} __attribute__ ((packed)) frame_type_t;
+
+// Struct which represent the http2 frame by its fields.
+// Checkout https://datatracker.ietf.org/doc/html/rfc7540#section-4.1 for frame format.
+struct http2_frame {
+    __u32 length : 24;
+    frame_type_t type;
+    __u8 flags;
+    __u8 reserved : 1;
+    __u32 stream_id : 31;
+} __attribute__ ((packed));
+
+#endif
