+++ pkg/network/ebpf/c/protocols/protocol-classification-common.h
@@ -1,0 +1,92 @@
+#ifndef __PROTOCOL_CLASSIFICATION_COMMON_H
+#define __PROTOCOL_CLASSIFICATION_COMMON_H
+
+#include <linux/types.h>
+
+#include "bpf_builtins.h"
+#include "bpf_telemetry.h"
+
+// Patch to support old kernels that don't contain bpf_skb_load_bytes, by adding a dummy implementation to bypass runtime compilation.
+#if LINUX_VERSION_CODE < KERNEL_VERSION(4, 5, 0)
+long bpf_skb_load_bytes_with_telemetry(const void *skb, u32 offset, void *to, u32 len) {return 0;}
+#endif
+
+#define CHECK_PRELIMINARY_BUFFER_CONDITIONS(buf, buf_size, min_buff_size)   \
+    if (buf_size < min_buff_size) {                                         \
+        return false;                                                       \
+    }                                                                       \
+                                                                            \
+    if (buf == NULL) {                                                      \
+        return false;                                                       \
+    }                                                                       \
+
+// Returns true if the packet is TCP.
+static __always_inline bool is_tcp(conn_tuple_t *tup) {
+    return tup->metadata & CONN_TYPE_TCP;
+}
+
+// Returns true if the payload is empty.
+static __always_inline bool is_payload_empty(struct __sk_buff *skb, skb_info_t *skb_info) {
+    return skb_info->data_off == skb->len;
+}
+
+// The method is used to read the data buffer from the __sk_buf struct. Similar implementation as `read_into_buffer_skb`
+// from http parsing, but uses a different constant (CLASSIFICATION_MAX_BUFFER).
+static __always_inline void read_into_buffer_for_classification(char *buffer, struct __sk_buff *skb, skb_info_t *info) {
+    u64 offset = (u64)info->data_off;
+
+#define BLK_SIZE (16)
+    const u32 len = CLASSIFICATION_MAX_BUFFER < (skb->len - (u32)offset) ? (u32)offset + CLASSIFICATION_MAX_BUFFER : skb->len;
+
+    unsigned i = 0;
+
+#pragma unroll(CLASSIFICATION_MAX_BUFFER / BLK_SIZE)
+    for (; i < (CLASSIFICATION_MAX_BUFFER / BLK_SIZE); i++) {
+        if (offset + BLK_SIZE - 1 >= len) { break; }
+
+        bpf_skb_load_bytes_with_telemetry(skb, offset, &buffer[i * BLK_SIZE], BLK_SIZE);
+        offset += BLK_SIZE;
+    }
+
+    // This part is very hard to write in a loop and unroll it.
+    // Indeed, mostly because of older kernel verifiers, we want to make sure the offset into the buffer is not
+    // stored on the stack, so that the verifier is able to verify that we're not doing out-of-bound on
+    // the stack.
+    // Basically, we should get a register from the code block above containing an fp relative address. As
+    // we are doing `buffer[0]` here, there is not dynamic computation on that said register after this,
+    // and thus the verifier is able to ensure that we are in-bound.
+    void *buf = &buffer[i * BLK_SIZE];
+    if (offset + 14 < len) {
+        bpf_skb_load_bytes_with_telemetry(skb, offset, buf, 15);
+    } else if (offset + 13 < len) {
+        bpf_skb_load_bytes_with_telemetry(skb, offset, buf, 14);
+    } else if (offset + 12 < len) {
+        bpf_skb_load_bytes_with_telemetry(skb, offset, buf, 13);
+    } else if (offset + 11 < len) {
+        bpf_skb_load_bytes_with_telemetry(skb, offset, buf, 12);
+    } else if (offset + 10 < len) {
+        bpf_skb_load_bytes_with_telemetry(skb, offset, buf, 11);
+    } else if (offset + 9 < len) {
+        bpf_skb_load_bytes_with_telemetry(skb, offset, buf, 10);
+    } else if (offset + 8 < len) {
+        bpf_skb_load_bytes_with_telemetry(skb, offset, buf, 9);
+    } else if (offset + 7 < len) {
+        bpf_skb_load_bytes_with_telemetry(skb, offset, buf, 8);
+    } else if (offset + 6 < len) {
+        bpf_skb_load_bytes_with_telemetry(skb, offset, buf, 7);
+    } else if (offset + 5 < len) {
+        bpf_skb_load_bytes_with_telemetry(skb, offset, buf, 6);
+    } else if (offset + 4 < len) {
+        bpf_skb_load_bytes_with_telemetry(skb, offset, buf, 5);
+    } else if (offset + 3 < len) {
+        bpf_skb_load_bytes_with_telemetry(skb, offset, buf, 4);
+    } else if (offset + 2 < len) {
+        bpf_skb_load_bytes_with_telemetry(skb, offset, buf, 3);
+    } else if (offset + 1 < len) {
+        bpf_skb_load_bytes_with_telemetry(skb, offset, buf, 2);
+    } else if (offset < len) {
+        bpf_skb_load_bytes_with_telemetry(skb, offset, buf, 1);
+    }
+}
+
+#endif
