+++ pkg/network/ebpf/c/protocols/protocol-classification-defs.h
@@ -4,20 +4,14 @@
 #include <linux/types.h>
 
 #include "amqp-defs.h"
+#include "http2-defs.h"
+#include "http-classification-defs.h"
 #include "mongo-defs.h"
+#include "redis-defs.h"
 
 // Represents the max buffer size required to classify protocols.
 // ATM, it is HTTP2_MARKER_SIZE.
 #define CLASSIFICATION_MAX_BUFFER (HTTP2_MARKER_SIZE)
-
-// Checkout https://datatracker.ietf.org/doc/html/rfc7540 under "HTTP/2 Connection Preface" section
-#define HTTP2_MARKER_SIZE 24
-
-// The minimal HTTP response has 17 characters: HTTP/1.1 200 OK\r\n
-// The minimal HTTP request has 16 characters: GET x HTTP/1.1\r\n
-#define HTTP_MIN_SIZE 16
-
-#define REDIS_MIN_FRAME_LENGTH 3
 
 // The enum below represents all different protocols we know to classify.
 // We set the size of the enum to be 8 bits, by adding max value (max uint8 which is 255) and
