+++ pkg/network/ebpf/c/protocols/redis-defs.h
@@ -1,0 +1,6 @@
+#ifndef __REDIS_DEFS_H
+#define __REDIS_DEFS_H
+
+#define REDIS_MIN_FRAME_LENGTH 3
+
+#endif
